+++ src/seacoral-cbmc/runner.ml
@@ -226,7 +226,7 @@ let cbmc_generic_process
     Log.debug "errors: `%a'" Sc_sys.File.print errors_file;
     Sc_sys.Lwt_file.descriptor errors_file [O_WRONLY; O_CREAT; O_TRUNC] 0o600
   in
-  let lines_stream_mbox = Lwt_mvar.create_empty () in
+  let output_lines, new_output_line = Lwt_stream.create () in
   let* proc =
     Sc_sys.Process.exec
       Sc_sys.Ezcmd.Std.(make "cbmc" |>
@@ -235,7 +235,7 @@ let cbmc_generic_process
                         rawf "-D%s" (str_of_mode mode) |>
                         to_cmd)
       ~stdin:(`FD_move (Lwt_unix.unix_file_descr inputs_fd))
-      ~stdout:(`Grab (MBox lines_stream_mbox))
+      ~stdout:(`Grab (Push_lines new_output_line))
       ~stderr:(`FD_move (Lwt_unix.unix_file_descr errors_fd))
       ~timeout
       ~on_success:(fun () -> Lwt.return_ok ())
@@ -244,7 +244,6 @@ let cbmc_generic_process
   let* cancel_kill =
     Sc_store.on_termination store ~h:(fun _ -> Sc_sys.Process.terminate proc)
   in
-  let* output_lines = Lwt_mvar.take lines_stream_mbox in
   Lwt.async begin fun () ->
     let oc = Lwt_io.of_fd outputs_fd ~mode:Output in
     let* () = Lwt_io.write_lines oc (Lwt_stream.clone output_lines) in